Revert "the builds no longer depend on a self built libindy but now pull from repo.sovrin.org"
This reverts commit f384ad6c1ad0de0859727e75f0d64a1cb87ba9ac.

+++ libsovtoken/build_scripts/ios/mac/mac.14.libsovtoken.build.sh
@@ -1,18 +1,5 @@
 #!/bin/sh
 
-set -xv
-
-function extract_target() {
-    case $1 in
-        aarch64-apple-ios) echo "arm64" ;;
-        armv7-apple-ios) echo "armv7" ;;
-        armv7s-apple-ios) echo "armv7s" ;;
-        i386-apple-ios) echo "i386" ;;
-        x86_64-apple-ios) echo "x86_64" ;;
-        \?) exit 1
-    esac
-}
-
 source ./shared.functions.sh
 
 START_DIR=${PWD}
@@ -20,16 +7,6 @@ WORK_DIR=$(get_work_dir)
 mkdir -p ${WORK_DIR}
 
 source ./mac.02.libindy.env.sh
-
-
-mkdir -p ${WORK_DIR}/sovtoken-indy-sdk
-if [ ! -f "${WORK_DIR}/sovtoken-indy-sdk/libindy.a" ] ; then
-    command pushd ${WORK_DIR}/sovtoken-indy-sdk > /dev/null
-    curl -sSLO ${LIBINDY_IOS_BUILD_URL}
-    tar -xf ${LIBINDY_FILE}
-    command popd > /dev/null
-    fi
-
 if [ ! -d "${WORK_DIR}/sovtoken-indy-sdk" ]; then
     echo STDERR "Unable to find ${WORK_DIR}/sovtoken-indy-sdk directory"
     exit 1
@@ -43,7 +20,6 @@ fi
 #########################################################################################################################
 # Now build libsovtoken
 #########################################################################################################################
-echo $(get_sovtoken_dir)
 cd $(get_sovtoken_dir)
 
 if [ "$DEBUG_SYMBOLS" = "debuginfo" ]; then
@@ -55,29 +31,33 @@ IFS=',()][' read -r -a targets <<<"${IOS_TARGETS}"
 echo "Building targets: ${targets[@]}"
 IFS="$bkpIFS"
 
-
 to_combine=""
 for target in ${targets[*]}
 do
-    LIBINDY=${WORK_DIR}/sovtoken-indy-sdk
-
-    export LIBINDY_DIR=${LIBINDY}/${target}
-    mkdir -p ${LIBINDY_DIR}
-    etarget=$(extract_target $target)
-
-    echo "LIBINDY_DIR=${LIBINDY_DIR}"
-    lipo -extract_family $etarget $LIBINDY/libindy.a -o $LIBINDY_DIR/libindy.a
+    export LIBINDY_DIR=${WORK_DIR}/sovtoken-indy-sdk/libindy/target/${target}/release
     cargo lipo --release --verbose --targets="${target}"
-    mv ./target/$target/release/libsovtoken.a ./target/$target/libsovtoken-unstripped.a
-    strip -S -x -o ./target/$target/libsovtoken.a -r ./target/$target/libsovtoken-unstripped.a
 
     to_combine="${to_combine} ./target/${target}/release/libsovtoken.a"
+
 done
 mkdir -p ./target/universal/release
 lipo -create $to_combine -o ./target/universal/release/libsovtoken.a
-cp ./target/universal/release/libsovtoken.a ./target/universal/libsovtoken-unstripped.a
-strip -S -x -o ./target/universal/libsovtoken.a -r ./target/universal/libsovtoken-unstripped.a
 
+for arch in ${IOS_TARGETS[@]}; do
+    if [ -f ./target.$arch/release/libsovtoken.a ] ; then
+        mv ./target/$arch/release/libsovtoken.a ./target/$arch/libsovtoken-unstripped.a
+        # to ensure the library is as small as possible....even though we are using cargo build /release
+        # we still have extraneous data in the library that needs to be removed
+        strip -S -x -o ./target/$arch/libsovtoken.a -r ./target/$arch/libsovtoken-unstripped.a
+    fi
+done
+
+if [ -f ./target/universal/release/libsovtoken.a ] ; then
+    cp ./target/universal/release/libsovtoken.a ./target/universal/libsovtoken-unstripped.a
+    # to ensure the library is as small as possible....even though we are using cargo build /release
+    # we still have extraneous data in the library that needs to be removed
+    strip -S -x -o ./target/universal/libsovtoken.a -r ./target/universal/libsovtoken-unstripped.a
+fi
 
 BUILD_TIME=$(date -u "+%Y%m%d%H%M")
 GIT_REV=$(git rev-parse --short HEAD)
